move `String#replaceAll` to stable ES
per Jun TC39 meeting
