+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Changelog
+##### Unreleased
+- `String#replaceAll` moved to stage 4, per Jun TC39 meeting
+
 ##### 3.6.5 - 2020.04.09
 - Updated Browserlist [#755](https://github.com/zloirock/core-js/issues/755)
 - Fixed `setImmediate` in Safari [#770](https://github.com/zloirock/core-js/issues/770), thanks [@dtinth](https://github.com/dtinth)
