+++ packages/core-js/es/string/replace-all.js
@@ -1,0 +1,4 @@
+require('../../modules/es.string.replace-all');
+var entryUnbind = require('../../internals/entry-unbind');
+
+module.exports = entryUnbind('String', 'replaceAll');
