+++ packages/core-js/features/string/replace-all.js
@@ -1,4 +1,1 @@
-require('../../modules/esnext.string.replace-all');
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('String', 'replaceAll');
+module.exports = require('../../es/string/replace-all');
