+++ packages/core-js/stable/string/replace-all.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/string/replace-all');
+
+module.exports = parent;
