+++ packages/core-js/stage/3.js
@@ -1,5 +1,4 @@
 require('../proposals/promise-any');
-require('../proposals/string-replace-all');
 var parent = require('./4');
 
 module.exports = parent;
