+++ packages/core-js/stage/4.js
@@ -1,6 +1,7 @@
 require('../proposals/global-this');
 require('../proposals/promise-all-settled');
 require('../proposals/string-match-all');
+require('../proposals/string-replace-all');
 var path = require('../internals/path');
 
 module.exports = path;
