+++ tests/pure/es.string.replace-all.js
