+++ tests/tests/es.string.replace-all.js
