+++ tests/tests/index.js
@@ -137,6 +137,7 @@ import './es.string.pad-start';
 import './es.string.raw';
 import './es.string.repeat';
 import './es.string.replace';
+import './es.string.replace-all';
 import './es.string.search';
 import './es.string.small';
 import './es.string.split';
@@ -294,7 +295,6 @@ import './esnext.set.symmetric-difference';
 import './esnext.set.union';
 import './esnext.string.at';
 import './esnext.string.code-points';
-import './esnext.string.replace-all';
 import './esnext.symbol.async-dispose';
 import './esnext.symbol.dispose';
 import './esnext.symbol.observable';
